Fixed Keep Policy for All 0's Keep policy didn't deal properly with setting all keep timeframes to 0.  Now, if all set to 0, then only the most recent backup/snapshot is kept, which seems like the reasonable expected behavior.
Also, for incremental backups, if keeps are available, then the most recent keep is used as the parent for the incremental.  We do this to create a nice chain of incremental backups.  If NO KEEPS, then the most recent snapshot/backup available is used, which could be something from the same day, if available.  Since no keeps, there is no chain so we might as well use the most recent available data we have.