First Addition of Hooks; Typos, Cosmetic The changes to btrbu-conf and README.md are cosmetic, fixing typos.
But in btrbu we now have hook roughed in.  Right now, they are only usable with a config file- no command line options to create hooks.  Only hooks related to snaps are done right now.  To make it useful, several substitutions were created so the user doesn't have to guess about filenames too much.  Also, had to add the `sudo` option so btrfs commands could be run by a normal user.

The snaphooks configuration is another table.  The keys match an archive name and the hook is loosely associated with that archive, however it is possible to specify other archives within a hook.  This was to facilitate use with my borgbackup situation, where I have different snapshots being sent to the same borg repository.  I create a wrapper script that invokes the appropriate borg commands to perform the backup, and I invoke this wrapper with the btrbu hook.

It might be better to make it a generalized hook? might be less confusing to other users...

Also, how many different hooks?  pre-snap? post snap?  post individual snap?  Same for backups?  lots of potential combos... What's useful?

Do i allow for some synchronization?  So that 1 hook doesn't get stepped on by a subsequent hook?  This is what my wrapper script takes care of, but if a synchronization was implemented then I wouldn't need wrapper scripts...

Like I said, it's roughed in...