Hooks Galore!
So decided on creating hooks for before snapshots, after snapshots and after each backup.  Configuration file options are defined as "presnaphooks", "postsnaphooks", "backuphooks".  Additionally, I've created some substitution strings that might be useful when crafting a hook.  The strings are framed in curly braces "{}" and btrbu will substitute an appropriate value for them.  Different substitution strings are available for different hooks, depending on what I thought might reasonably be needed for each type of hook.

Presnaphooks is a list of command to run.  These will all be run prior to any snapshots are performed.  Substitutions allowed are {timestamp}, {snapshotdir}, and {backupdir}.

Postsnaphooks are started after snapshots are done.  They are intended to allow a user to work on the snapshots just taken.  The configuration is a table with the archive name as a key and the command as the value. Substitutions allowed are {snapshot}, {archive}, {timestamp}, {snapshotdir}, {backupdir}.

Backup hooks are started after each backup is completed.  Backup hooks are also configured using a table with the archive name as the key and the command as the value.  Substitutions allowed are {backup}, {snapshot}, {archive}, {timestamp}, {snapshotdir}, {backupdir}.